Fix aliasing error for continuous attributes
Current inplementation creates a list of aliases to the same list - when features are in wildly different ranges causes the range of every feature to be the range of the largest feature. This change removes the aliasing error, while also removing the error of not being able to have a minimum in the range that is greater than 0, or a maximum less than 0.